Use half float textures
This allows for the simulation to be ran on more devices, expecially as iOS devices apparently cannot render to floating point textures, but they can render to half float textures.
No loss in quality is detected, but some inprecision can be expected as the simulation is technically less precise.
Half float textures has been implemented for both webgl and webgl2.