iOS (#10)
* Update actions
* `bump.yml` should only trigger on non version updating pushes
* `build.yml` should cache `node_modules`

* Format option in textures

* Use intermediary texture to render

* GUI tweaking

* Update build.yml

Small oversight in caching. I think this fixes it

* Revert "Use intermediary texture to render"

This reverts commit 3789565ac7b4c6f6041ccb685233208f9d0c7c18.

It was discovered that the main issue was not rendering to a floating point texture, instead the issue was that it was a floating point texture and not a half float texture...
By reverting this commit a solution to instead use a half float textures will be used

* Revert "Format option in textures"

This reverts commit b9e6fb950bd60f46fc4f9ea106164b51e416eef5.

See commit b61e604

* New default params

* Use half float textures

This allows for the simulation to be ran on more devices, expecially as iOS devices apparently cannot render to floating point textures, but they can render to half float textures.
No loss in quality is detected, but some inprecision can be expected as the simulation is technically less precise.
Half float textures has been implemented for both webgl and webgl2.